requirements.txt
fastapi
uvicorn
openai
python-dotenv
pytest
httpx
sqlalchemy